+++ js/data.js
@@ -1,3 +1,5 @@
+import {getNumberFromInterval} from './util.js';
+
 //Вводные данные
 const PHOTO_COUNT = 25;
 const MIN_LIKE = 15;//Минимальное количество лайков
@@ -27,4 +29,34 @@ const MESSAGES = [
   'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
 ];
 
-export {PHOTO_COUNT, MIN_LIKE, MAX_LIKE, DESCRIPTIONS, NAMES, MESSAGES};
+const arrayOfPhotoObjects = [];//Итоговый массив объектов
+
+//Тут получаю массив объектов-комментов
+function getComments(countOfComments) {
+  const comments = [];//Массив для комментов
+  for(let j = 0; j < countOfComments; j++) {
+    comments[j] = {
+      id: j,
+      avatar: `img/avatar${getNumberFromInterval(1, 6)}.svg`,
+      message: MESSAGES[getNumberFromInterval(0, MESSAGES.length-1)],
+      name: NAMES[getNumberFromInterval(0, NAMES.length - 1)]
+    };
+  }
+  return comments;
+}
+
+//Тут получаю итоговый массив объектов
+const createPhotoObjects = function(counter) {
+  for(let i = 1; i <= counter; i++) {
+    arrayOfPhotoObjects[i] = {
+      //id: i,
+      url: `photos/${i}.jpg`,
+      //description: DESCRIPTIONS[getNumberFromInterval(0, DESCRIPTIONS.length-1)],
+      likes: getNumberFromInterval(MIN_LIKE, MAX_LIKE),
+      comment: getComments(getNumberFromInterval(1,5))
+    };
+  }
+  return arrayOfPhotoObjects;
+}
+
+export { createPhotoObjects, PHOTO_COUNT }
